next latest is not compatible with react 17
I started learning next js today and quickly became stuck by having the recommended command fail due to a dependency mismatch. Since next already has react and react-dom as a peer dependency, by deleting the explicit reference to react and react-dom I was able to continue with the tutorial. A different solution would be to explicitly peg react 18, but I worry a similar breakage would occur in the future.